Add things in controller

+++ classified-offering/Controllers/ClassifiedOfferingsController.cs
@@ -14,11 +14,28 @@ namespace classified_offering.Controllers
     {
         private ClassifiedOfferingDbContext db = new ClassifiedOfferingDbContext();
 
+        private ActionResult CheckAuthenticated(int? id = null)
+        {
+            User user = (User)Session["connectedUser"];
+            return user == null || (id != null && user.ID != id && user.Role != 0)
+                 ? Redirect("/User/SignIn")
+                 : null;
+        }
+
         // GET: ClassifiedOfferings
         public ActionResult Index()
         {
-            var classifiedOfferings = db.ClassifiedOfferings.Include(c => c.Creator);
-            return View(classifiedOfferings.ToList());
+            ActionResult NotAuthenticated = CheckAuthenticated();
+            if (NotAuthenticated != null)
+            {
+                return NotAuthenticated;
+            }
+            User connectedUser = (User) Session["connectedUser"];
+            var classifiedOfferings = connectedUser.Role == 0
+                ? db.ClassifiedOfferings.ToList()
+                : db.ClassifiedOfferings.Where(c => c.CreatorID == connectedUser.ID).ToList();
+            
+            return View(classifiedOfferings);
         }
 
         // GET: ClassifiedOfferings/Details/5
@@ -28,7 +45,7 @@ namespace classified_offering.Controllers
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
             }
-            ClassifiedOffering classifiedOffering = db.ClassifiedOfferings.Find(id);
+            ClassifiedOffering classifiedOffering = db.ClassifiedOfferings.Include(co => co.Participations).First(co => co.ID == id);
             if (classifiedOffering == null)
             {
                 return HttpNotFound();
@@ -39,7 +56,11 @@ namespace classified_offering.Controllers
         // GET: ClassifiedOfferings/Create
         public ActionResult Create()
         {
-            ViewBag.CreatorID = new SelectList(db.Users, "ID", "Pseudo");
+            ActionResult NotAuthenticated = CheckAuthenticated();
+            if (NotAuthenticated != null)
+            {
+                return NotAuthenticated;
+            }
             return View();
         }
 
@@ -50,30 +71,43 @@ namespace classified_offering.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Create([Bind(Include = "ID,Name,CreationDate,CreatorID")] ClassifiedOffering classifiedOffering)
         {
+            ActionResult NotAuthenticated = CheckAuthenticated();
+            if (NotAuthenticated != null)
+            {
+                return NotAuthenticated;
+            }
+
             if (ModelState.IsValid)
             {
+                classifiedOffering.CreatorID = ((User) Session["connectedUser"]).ID;
                 db.ClassifiedOfferings.Add(classifiedOffering);
                 db.SaveChanges();
                 return RedirectToAction("Index");
             }
-
-            ViewBag.CreatorID = new SelectList(db.Users, "ID", "Pseudo", classifiedOffering.CreatorID);
             return View(classifiedOffering);
         }
 
         // GET: ClassifiedOfferings/Edit/5
         public ActionResult Edit(int? id)
         {
+           
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
             }
             ClassifiedOffering classifiedOffering = db.ClassifiedOfferings.Find(id);
+
             if (classifiedOffering == null)
             {
                 return HttpNotFound();
             }
-            ViewBag.CreatorID = new SelectList(db.Users, "ID", "Pseudo", classifiedOffering.CreatorID);
+
+            ActionResult NotAuthenticated = CheckAuthenticated(classifiedOffering.CreatorID);
+            if (NotAuthenticated != null)
+            {
+                return NotAuthenticated;
+            }
+            
             return View(classifiedOffering);
         }
 
@@ -84,13 +118,17 @@ namespace classified_offering.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Edit([Bind(Include = "ID,Name,CreationDate,CreatorID")] ClassifiedOffering classifiedOffering)
         {
+            ActionResult NotAuthenticated = CheckAuthenticated(classifiedOffering.CreatorID);
+            if (NotAuthenticated != null)
+            {
+                return NotAuthenticated;
+            }
             if (ModelState.IsValid)
             {
                 db.Entry(classifiedOffering).State = EntityState.Modified;
                 db.SaveChanges();
                 return RedirectToAction("Index");
             }
-            ViewBag.CreatorID = new SelectList(db.Users, "ID", "Pseudo", classifiedOffering.CreatorID);
             return View(classifiedOffering);
         }
 
@@ -106,6 +144,11 @@ namespace classified_offering.Controllers
             {
                 return HttpNotFound();
             }
+            ActionResult NotAuthenticated = CheckAuthenticated(classifiedOffering.CreatorID);
+            if (NotAuthenticated != null)
+            {
+                return NotAuthenticated;
+            }
             return View(classifiedOffering);
         }
 
@@ -114,7 +157,17 @@ namespace classified_offering.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult DeleteConfirmed(int id)
         {
+            ActionResult NotAuthenticated = CheckAuthenticated();
+            if (NotAuthenticated != null)
+            {
+                return NotAuthenticated;
+            }
             ClassifiedOffering classifiedOffering = db.ClassifiedOfferings.Find(id);
+            User connectedUser = (User) Session["connectedUser"];
+            if (connectedUser.ID != classifiedOffering.CreatorID || connectedUser.Role != 0)
+            {
+                return NotAuthenticated;
+            }
             db.ClassifiedOfferings.Remove(classifiedOffering);
             db.SaveChanges();
             return RedirectToAction("Index");
